% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_age_bins}
\alias{get_age_bins}
\title{Get age bins for use in the model}
\usage{
get_age_bins()
}
\value{
A data frame containing the labels for the age bins, and their
start and end values
}
\description{
A function that gets the age bins used in the model.
}
