% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_compartment_indices}
\alias{get_compartment_indices}
\title{Get compartment indices used throughout the model}
\usage{
get_compartment_indices()
}
\value{
A list containing entries: \code{dim} giving the dimensions of each
compartment, currently: group = 18, vax = 4;
\code{group}, a named list of the array indices corresponding to the
first dimension of model compartments: 16 5-year age bands + SW + PBS;
and \code{vax}, a named list of the vaccine strata used in
the second dimension of model compartments:
\enumerate{
\item historic smallpox; 2. unvaccinated; 3. one-dose; 4. two-dose.
}

Use this function wherever you need to refer to these standards in your code
to avoid duplication, and make modifying universally easier.
}
\description{
A function that gets the compartment indices used in the model
}
