% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{mpoxseir_date}
\alias{mpoxseir_date}
\alias{mpoxseir_date_as_date}
\alias{as_mpoxseir_date}
\alias{as_date}
\title{Date handling for mpoxseir}
\usage{
mpoxseir_date(date)

mpoxseir_date_as_date(date)

as_mpoxseir_date(date)

as_date(date)
}
\arguments{
\item{date}{A Date object, or something that can be converted to
one, or an "mpoxseir date"; see Details}
}
\value{
An integer, being the number of days into 2023
}
\description{
We need to map "dates" onto \code{\link[dust:dust]{dust::dust}}'s concept of model
"step" and we do this by mapping a date such as \code{2024-03-02} into
the number of days since the start of 2023 (426 here, with the 1st of
January being day 0). We call this integer number an "mpoxseir date".
}
\details{
There are several related functions here
\itemize{
\item \code{mpoxseir_date} converts its argument into an R \code{Date} object,
then applies this tranformation. If the argument is not a \code{Date}
object or a string representing one, an error will be thrown.
\item \code{mpoxseir_date_to_date} does the reverse conversion to
\code{mpoxseir_date}, converting an integer mpoxseir date into an R
\code{Date}
\item \code{as_mpoxseir_date} does the same conversion as \code{mpoxseir_date}
but will assume that an integer \emph{already} represents an mpoxseir
date and will return it unmodified rather than erroring.
\item \code{as_date} does a string to date conversion, using \code{\link[=as.Date]{as.Date()}}
but requiring the dates are in ISO 8601 (YYYY-MM-DD) format (it
is a helper that avoids conversion to \code{NA}, instead throwing an
error)
}
}
\examples{
# Convert dates into mpoxseir dates:
mpoxseir::mpoxseir_date("2023-01-01")
mpoxseir::mpoxseir_date(c("2024-03-01", "2024-10-01"))

# Reverse the conversion:
mpoxseir::mpoxseir_date_as_date(0)
mpoxseir::mpoxseir_date_as_date(c(425, 639))

# Double conversion not possible with mpoxseir_date...
try(mpoxseir::mpoxseir_date(61))
# ...but allowed with as_mpoxseir_date
mpoxseir::as_mpoxseir_date(61)

# Strict date conversion with as_date
mpoxseir::as_date("2024-03-01")
try(mpoxseir::as_date("03-01-2024"))
}
