% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters_demographic}
\alias{parameters_demographic}
\title{Get demographic parameters}
\usage{
parameters_demographic(
  region,
  mixing_matrix = "Zimbabwe",
  p_SW = NULL,
  p_HCW = NULL
)
}
\arguments{
\item{region}{The region for the parameters, must be either \code{"equateur"},
\code{"sudkivu"}, \code{"burundi"} or \code{"bujumbura"}}

\item{mixing_matrix}{The mixing matrix must be either \code{"Zimbabwe"},
\code{"synthetic_home"}, or \code{"synthetic_all"}}

\item{p_SW}{The proportion of SW-age groups that are sex workers. Note that
e.g. a value of 0.01 means 1\% of all SW-age groups are sex workers and not
just 1\% of women in those groups. Default is NULL, in which case we use
the default value for the region given in the package}

\item{p_HCW}{The proportion of HCW-age groups that are healthcare workers.
Default is NULL, in which case we use the default value for the region
given in the package}
}
\value{
A list containing all the demographic parameters
}
\description{
A function that returns the demographic parameters for use in the model
}
