% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{proportion_in_age_bins}
\alias{proportion_in_age_bins}
\title{calculate the proportion of each age group that lies between \code{min_age}
and \code{max_age} (inclusive) based on uniform distribution within each age band}
\usage{
proportion_in_age_bins(min_age, max_age, bins = NULL)
}
\arguments{
\item{min_age}{a scalar giving the bottom of the age range}

\item{max_age}{a scalar giving the top of the age range}
}
\value{
a vector of length n_age = 16 giving the proportion for each age band
}
\description{
A function that calculates the proportion of each age group than lies inside
a given interval
}
