% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_mpoxSEIR_targetedVax_single}
\alias{run_mpoxSEIR_targetedVax_single}
\title{Run the model with single set of parameter values}
\usage{
run_mpoxSEIR_targetedVax_single(
  region,
  initial_infections,
  n_weeks,
  R0_hh,
  R0_sw_st,
  beta_z_max,
  n_vax = 0,
  daily_doses = 0,
  N_prioritisation_steps = 0,
  prioritisation_strategy = 0,
  vaccination_coverage_target = 0,
  vaccine_uptake = 0,
  ve_T = 0,
  ve_I = 0,
  ve_D = 0,
  vaccination_campaign_length = 0,
  overrides = list(),
  n_particles = 1,
  n_threads = 1,
  seed = 42,
  deterministic = TRUE,
  outputs_retained = NULL
)
}
\arguments{
\item{region}{The region to run the model for, must be either \code{"equateur"}
or \code{"sudkivu"}}

\item{initial_infections}{The initial number of infections to seed with}

\item{n_weeks}{number of weeks to run for}

\item{R0_hh}{R0 for the household}

\item{R0_sw_st}{R0 for sex workers to people who buy sex}

\item{beta_z_max}{beta for the age-group with highest zoonotic
transmission (a number)}

\item{n_vax}{number of vaccination compartments (integer, basis for an
additional dimension in odin states)}

\item{daily_doses}{the daily number of doses administered (matrix of
vaccination_campaign_length * number of vaccination compartments)}

\item{N_prioritisation_steps}{the number of different vaccination
prioritisation categories we're considering}

\item{prioritisation_strategy}{what each step corresponds to in terms of
strategy (matrix of n_group * N_prioritisation_steps, with 1s and 0s
indicating whether a group is included in prioritisation step)}

\item{vaccination_coverage_target}{vaccination coverage target for each
group and prioritisation step (matrix of n_group * N_prioritisation_steps)}

\item{vaccine_uptake}{max achievable coverage for each group (vector of
length n_group)}

\item{ve_T}{vaccine efficacy against onwards transmissibility for each
vaccinated compartment (vector of length n_vax)}

\item{ve_I}{vaccine efficacy against infection for each vaccinated
compartment (vector of length n_vax)}

\item{ve_D}{vaccine efficacy against death for each vaccinated
compartment (vector of length n_vax)}

\item{vaccination_campaign_length}{length of the vaccination campaign
(in timesteps NOT days - CHECK WITH RUTH THIS IS RIGHT)}

\item{overrides}{list of other model parameters which if specified will
overwrite the defaults}

\item{n_particles}{Number of particles}

\item{n_threads}{Number of threads}

\item{seed}{The random seed to use}

\item{deterministic}{Logical, whether to run the model deterministically or
not}

\item{outputs_retained}{The outputs to retain}
}
\value{
The output from running the model
}
\description{
Run the model with single set of parameter values
}
