% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters_demographic}
\alias{parameters_demographic}
\title{Get demographic parameters}
\usage{
parameters_demographic()
}
\value{
A list containing all the demographic parameters
}
\description{
A function that returns the demographic parameters for use in the model
}
