% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters_fixed}
\alias{parameters_fixed}
\title{Get fixed parameters for use in the model}
\usage{
parameters_fixed(
  region,
  initial_infections,
  use_ve_D = FALSE,
  mixing_matrix = "Zimbabwe",
  p_SW = NULL,
  overrides = list()
)
}
\arguments{
\item{region}{The region for the parameters, must be either \code{"equateur"},
\code{"sudkivu"}, \code{"burundi"} or \code{"bujumbura"}}

\item{initial_infections}{The initial number of infections}

\item{use_ve_D}{logical, indicating whether model should allow for vaccine
efficacy against death (above and beyond protection against infection)}

\item{mixing_matrix}{The mixing matrix must be either \code{"Zimbabwe"},
\code{"synthetic_home"}, or \code{"synthetic_all"}}

\item{p_SW}{The proportion of SW-age groups that are sex workers. Note that
e.g. a value of 0.01 means 1\% of all SW-age groups are sex workers and not
just 1\% of women in those groups. Default is NULL, in which case we use
the default value for the region given in the package}

\item{overrides}{A list, containing any parameters for which you want to
override the default values}
}
\value{
A list of the fixed parameters
}
\description{
A function that gets the fixed parameters for use in the model
}
