% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters_fixed}
\alias{parameters_fixed}
\title{Get fixed parameters for use in the model}
\usage{
parameters_fixed(region, initial_infections, overrides = list())
}
\arguments{
\item{region}{The region for the parameters, must be either \code{"equateur"} or
\code{"sudkivu"}}

\item{initial_infections}{The initial number of infections}

\item{overrides}{A list, containing any parameters for which you want to
override the default values}
}
\value{
A list of the fixed parameters
}
\description{
A function that gets the fixed parameters for use in the model
}
